+++ lifelines/fitters/generalized_gamma_fitter.py
@@ -101,7 +101,7 @@ class GeneralizedGammaFitter(KnownModelParametericUnivariateFitter):
     _bounds = [(None, None), (None, None), (None, None)]
     _compare_to_values = np.array([0, 0, 1])
 
-    def _get_initial_values(self, Ts, E, *args):
+    def _create_initial_point(self, Ts, E, *args):
         if CensoringType.is_right_censoring(self):
             log_data = log(Ts[0])
         elif CensoringType.is_left_censoring(self):