+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ##### New features
  - Improvements to the __repr__ of models that takes into accounts weights.
+ - Better support for predicting on Pandas Series
 
 ##### Bug fixes
  - Fixed issue where `fit_interval_censoring` wouldn't accept lists.
@@ -11,6 +12,7 @@
 
 ##### API Changes
  - `_get_initial_value` in parametric univariate models is renamed `_create_initial_point`
+
 
 #### 0.22.4 - 2019-09-04